backend/requirements.txt
Django>=4.0,<5.0
psycopg2>=2.8
djangorestframework>=3.0,<4.0